Updates made in response to approver comments: file name changes, deletion of superfluous code, and refactoring functionality to be more generic.
